Correctly scope private methods in Sidewinder
The `private` keyword in the regular class scope declares _instance_
methods to be private, but does not apply to class methods. Use the
`class << self` construct to open the singleton class scope and declare
private instance methods there. (This is the scope that is actually,
magically opened when using `def self.some_class_method`.)

An alternative would be to use `Class#private_class_method`.